+++ src/components/checkbox/Checkbox.vue
@@ -10,6 +10,7 @@
             v-model="newValue"
             type="checkbox"
             :disabled="disabled"
+            :required="required"
             :name="name"
             :value="nativeValue"
             :true-value="trueValue"
@@ -27,6 +28,7 @@
             nativeValue: [String, Number, Boolean, Function, Object, Array, Symbol],
             type: String,
             disabled: Boolean,
+            required: Boolean,
             name: String,
             size: String,
             trueValue: {